Neighborhoods from Geocoded Locations
This is something I did last week. Every address can be associated with a planning neighborhood in the city using the lat/long boundaries of the neighborhoods. Might not work this into the final report but something useful to have: a function to determine if points fall within complex geometry boundaries.